training/requirements.txt
torch
torchvision
python-dotenv
ultralytics
mlflow
pyfiglet
pytest
boto3
optuna